backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = ""
authors = [
    {name = "Jean Torre",email = "jean.torre21@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uuid (>=1.30,<2.0)",
    "pytz (>=2025.2,<2026.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
